third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-scrollBoundaryBehavior-serialization.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<!-- https://wicg.github.io/scroll-boundary-behavior/#scroll-boundary-behavior-properties -->

<style>
  #container {
    scroll-boundary-behavior-x: none;
    scroll-boundary-behavior-y: contain;
  }
  #child {
    scroll-boundary-behavior: inherit;
  }
</style>
<div id="container">
  <div id="child"></div>
</div>
<div id="target"></div>

<script>
'use strict';

test(function() {
  assert_equals(getComputedStyle(document.body).scrollBoundaryBehavior, 'auto auto');
  assert_equals(getComputedStyle(container).scrollBoundaryBehavior, 'none contain');
  assert_equals(getComputedStyle(child).scrollBoundaryBehavior, 'none contain');
}, "Test that getComputedStyle serializes scroll-boundary-behavior-x and scroll-boundary-behavior-y");

var values = [
  // single keyword
  {input: 'contain', expected: 'contain contain'},
  {input: 'none', expected: 'none none'},
  {input: 'auto', expected: 'auto auto'},

  // two keywords
  {input: 'contain none', expected: 'contain none'},
  {input: 'none auto', expected: 'none auto'},
  {input: 'auto contain', expected: 'auto contain'},
];

test(function() {
  for (var value of values) {
    target.style.scrollBoundaryBehavior = value.input;
    assert_equals(getComputedStyle(target).scrollBoundaryBehavior, value.expected);
  }
}, "Test that getComputedStyle serializes scrollBoundaryBehavior as per the spec");
</script>
